arxiv_crawl_v2/requirements.txt
arxiv
PyMuPDF
scikit-learn
nltk
pyyaml
sentence-transformers
keybert